+++ src/dawn_wire/WireClient.cpp
@@ -25,7 +25,7 @@ namespace dawn_wire {
         mImpl.reset();
     }
 
-    const volatile char* WireClient::HandleCommands(const volatile char* commands, size_t size) {
+    const volatile char* WireClient::HandleCommands(const volatile char* commands, uint32_t size) {
         return mImpl->HandleCommands(commands, size);
     }
